# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SchoolPilot API
  description: |
    Multi-tenant school management SaaS API.
    Authenticated routes require `Authorization: Bearer <accessToken>`.
    OpenAPI spec also served at `GET /api/docs/openapi.yaml`.
  version: 1.0.0
  contact:
    name: Great Goodluck Schools
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.yourschoolpilot.com
    description: Production
...
paths:
...
  /api/schools/{id}/subscription/checkout:
    post:
      tags: [SaaS]
      summary: Start Paystack subscription checkout
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [planSlug]
              properties:
                planSlug: { type: string }
                billingInterval: { type: string, enum: [monthly, quarterly, yearly] }
                couponCode: { type: string }
      responses:
        '200':
          description: checkoutUrl and reference
  /api/schools/{id}/subscription/cancel:
    post:
      tags: [SaaS]
      summary: Cancel school subscription
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cancelled
  /api/subscription/verify:
    post:
      tags: [SaaS]
      summary: Verify Paystack payment after checkout
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reference]
              properties:
                reference: { type: string }
      responses:
        '200':
          description: Subscription activated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT